% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileprocessing.R
\name{rbioIMP}
\alias{rbioIMP}
\title{rbioIMP}
\usage{
rbioIMP(fileName, method = "mean", transpo = TRUE)
}
\arguments{
\item{fileName}{Inpule file name. Case sensitive and be sure to type with quotation marks. Currently only takes \code{.csv} files.}

\item{method}{Imputation method. Options are \code{"mean"}, \code{"random"} and \code{"regression"}. Default is \code{"mean"}.}

\item{transpo}{If the output file is transposed. Default is \code{TRUE}.}
}
\value{
Outputs a \code{dataframe} object with all the missing value imputed.
}
\description{
data imputation for raw data.
}
\details{
Make sure to make the input the data with only the tagert variable columns, meaning no annotation or index variables. The format would be target variables only (column). \code{transpo} arugment is particularly useful for the upcoming data normalization and random forest operations.
}
\examples{
\dontrun{
mydfm <- rbioIMP("mydata.csv", method = "mean") # make sure no aannotation variable present in the data file.
}
}

