% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileprocessing.R
\name{rbioIMP}
\alias{rbioIMP}
\title{rbioIMP}
\usage{
rbioIMP(dfm, fct, annot, method = "mean", transpo = TRUE)
}
\arguments{
\item{dfm}{Input dataframe. Case sensitive and be sure to type with quotation marks.}

\item{fct}{Input factor variable, based on the which the missing data is imputed.}

\item{annot}{Input variable or character string used for variable name for the transposed dataframe.}

\item{method}{Imputation method. Options are \code{"mean"}, \code{"random"} and \code{"regression"}. Default is \code{"mean"}.}

\item{transpo}{If the output file is transposed. Default is \code{TRUE}.}
}
\value{
Outputs a \code{dataframe} object with all the missing value imputed.
}
\description{
data imputation for raw data based on the factor.
}
\details{
Make sure to make the input the data with only the tagert variable columns, meaning no annotation or index variables. The format would be target variables only (column). \code{transpo} arugment is particularly useful for the upcoming data normalization and random forest operations.
}
\examples{
\dontrun{
mydfm <- rbioIMP(raw[-c(1:2)], raw$Conditions, raw$sampleName, method = "mean", transpo = TRUE) # make sure no annotation variable present in the data file.
}
}

