% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbioRF_iterOOB.R
\name{rbioRF_iterOOB}
\alias{rbioRF_iterOOB}
\title{rbioRF_iterOOB}
\usage{
rbioRF_iterOOB(x, targetVar, nTimes = 50, nTree = 1001, multicore = TRUE)
}
\arguments{
\item{x}{Input dataframe or matrix. Make sure to arrange the data with features as column names.}

\item{targetVar}{The target variable for random forest feature selection. This is a factor object.}

\item{nTimes}{Number of iteration of random forest vi computation. Default is \code{50} times.}

\item{nTree}{Number of trees generated for each random forest iteration. Default is \code{1001} trees.}

\item{multicore}{If to use parallel computing. Default is \code{TRUE}.}
}
\value{
Outputs a \code{matrix} object with  OOB error rates.
}
\description{
Iterative nested random froest variable importance (vi) and OOB error rate computation. (to be completed)
}
\details{
Make sure to arrange data (dfm) with feature (e.g., gene) as variables (i.e., columns), and rownames as sample names.
}
\examples{
\dontrun{
rbioRF_iterOOB(training_HCvTC, tgtVar_HCvTC, multicore = TRUE)
}
}

