% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbioRF.R
\name{rbioRF_vi}
\alias{rbioRF_vi}
\title{rbioRF_vi}
\usage{
rbioRF_vi(dfm, targetVar, nTimes = 50, transpo = TRUE, nTree = 1001)
}
\arguments{
\item{dfm}{Input data frame.}

\item{targetVar}{The target variable for random forest feature selection.}

\item{nTimes}{Number of iteration of random forest vi computation. Default is \code{50} times.}

\item{transpo}{If the dataframe needs to be transposed before random forest. Default is \code{TRUE}.}

\item{nTree}{Number of trees generated for each random forest iteration. Default is \code{1001} trees.}
}
\value{
Outputs a \code{matrix} object with vi values for all the random forest iterations for each feature.
}
\description{
Iterative random froest variable importance (vi) computation.
}
\examples{
\dontrun{
rbioRF(dataframe, tgtvar_CntlvsStress, transpo = FLASE, nTree = 501)
}
}

