% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbioRF.R
\name{rbioRF_viplot}
\alias{rbioRF_viplot}
\title{rbioRF_viplot}
\usage{
rbioRF_viplot(fsdfm, n = "all", Title = NULL,
  xLabel = "Mean Decrease in Accuracy", yLabel = NULL, errorbar = "SEM",
  errorbarWidth = 0.2, xTxtSize = 10, yTxtSize = 10, plotWidth = 170,
  plotHeight = 150)
}
\arguments{
\item{fsdfm}{Input data frame.}

\item{n}{Number of features to show. Takes integer numbers. Default is \code{"all"} (make sure to include quotation marks).}

\item{Title}{Figure title. Make sure to use quotation marks. Use \code{NULL} to hide. Default is \code{NULL}.}

\item{xLabel}{X-axis label. Make sure to use quotation marks. Use \code{NULL} to hide. Default is \code{NULL}.}

\item{yLabel}{Y-axis label. Make sure to use quotatio marks. Use \code{NULL} to hide. Default is \code{"Mean Decrease in Accurac"}}

\item{errorbar}{The type of errorbar in the graph. Options are \code{"SEM"} (standard error of the mean) or \code{"SD"} (standard deviation). Default is \code{"SEM"}.}

\item{errorbarWidth}{The width of the errorbar. Default is \code{0.2}.}

\item{xTxtSize}{Font size for the x-axis text. Default is \code{10}.}

\item{yTxtSize}{Font size for the y-axis text. Default is \code{10}.}

\item{plotWidth}{The width of the figure for the final output figure file. Default is \code{170}.}

\item{plotHeight}{The height of the figure for the final output figure file. Default is \code{150}.}
}
\value{
Outputs a \code{pdf} figure file.
}
\description{
Generate bargraph for vi values obtained from itarative random forest
}
\examples{
\dontrun{
rbioRF(dataframe, tgtvar_CntlvsStress, transpo = FLASE, nTree = 501)
}
}

